Implement skills for every character

+++ src/main/java/dev/spiritworker/util/Utils.java
@@ -19,6 +19,6 @@ public class Utils {
 	}
 
 	public static int getAppearanceType(int appearance) {
-		return (int) Math.floor((float) appearance / 1000f);
+		return (int) Math.floor(appearance / 1000f);
 	}
 }
